Update and rename README.txt to README.md
I always disliked that I left this as a *.txt file when it should have been a *.md. Fixed the formatting some too. This engine has a ton of issues with not only architecture but coding style as well. Maybe in the future I shall revisit it under a new name.